Refactoring & JWT Permissions Claims

+++ Application/Interfaces/Repositories/IRepository.cs
@@ -6,8 +6,8 @@ namespace Application.Interfaces.Repositories
 {
     public interface IRepository<TEntity> where TEntity : class
     {
-        Task<IEnumerable<TEntity>> GetAllAsync();
-        Task<TEntity> GetByIdAsync(long id);
+        Task<IEnumerable<TEntity>> GetAllAsync();   
+        Task<TEntity> GetByIdAsync(int id);
         Task Add(TEntity entity);
         void Update(TEntity entity);
         void Delete(TEntity entity);
